C++ compiler warning
git-svn-id: svn://www.xmlblaster.org/xmlBlaster/trunk@17542 00fa6119-6a0f-0410-b6da-8b96a2b62773

+++ testsuite/src/c/TestStress.c
@@ -160,7 +160,7 @@ static const char * test_stress()
    mu_assert("No update arrived", *updateContent != '\0');
    if ((int)updateCounter < numPublish) {
       freeXmlBlasterAccessUnparsed(xa);
-      mu_assert("Missing updates", updateCounter == numPublish);
+      mu_assert("Missing updates", (int)updateCounter == numPublish);
    }
    else if ((int)updateCounter > numPublish) {
       printf("[client] WARN: Publish of %d messages but received %d updates\n", numPublish, (int32_t)updateCounter);
